bf16 prepare dataset fix

+++ training/prepare_dataset.py
@@ -287,12 +287,12 @@ to_pil_image = transforms.ToPILImage(mode="RGB")
 
 
 def save_image(image: torch.Tensor, path: pathlib.Path) -> None:
-    image = to_pil_image(image)
+    image = to_pil_image(image.float())
     image.save(path)
 
 
 def save_video(video: torch.Tensor, path: pathlib.Path, fps: int = 8) -> None:
-    video = [to_pil_image(frame) for frame in video]
+    video = [to_pil_image(frame) for frame in video.float()]
     export_to_video(video, path, fps=fps)
